Updated the AI Ground Eater a bit.
Changed priority to 0, because this is based on items that can decompose
on the ground, animals should try to eat them more then run away and
swim. Which is pretty rare they would be doing both.

switched to use the isInLove method.
changed the inLove timer to 2400 instead of 600, this allows more time
for other AI to tick.
simplified the logic of the eat due to the fact that it no longer needs
to return anything.